Add a pages folder and move pages to that folder.
