+++ lib/widgets/carouselDisplayWidget.dart
@@ -3,7 +3,7 @@ import 'package:cached_network_image/cached_network_image.dart';
 
 import 'package:readr_app/helpers/constants.dart';
 import 'package:readr_app/models/record.dart';
-import 'package:readr_app/storyPage.dart';
+import 'package:readr_app/pages/storyPage.dart';
 
 class CarouselDisplayWidget extends StatelessWidget {
   final Record record;
